:recycle: refactor : auth 관련 페이지 로그인 상태 로직 변경
